refactor(tests): split test_episodic_memory.py into focused modules
The 1927-line monolith is replaced by six files each under 500 lines,
satisfying the file-size policy enforced by scripts/checks/file_size.py.
Shared fixtures (memory, memory_pair) are promoted to conftest.py so
they are available to all modules without duplication.

Split:
- test_episodic_memory_core.py       — dataclass, migrations, CRUD,
                                       access tracking, agent isolation,
                                       edge cases, MemoryLifecycle
- test_episodic_memory_fts5.py       — FTS5 availability, trigger sync,
                                       ranking, malformed-query fallback,
                                       LIKE escaping, context retrieval,
                                       query sanitisation
- test_episodic_memory_validation.py — importance/summary/limit guards,
                                       zero-importance recall, ln() check,
                                       SQL score syntax, BM25 normalisation,
                                       min_score range validation, defaults
- test_episodic_memory_compression.py — LLM-driven summarisation, batch
                                        processing, partial failures,
                                        agent scoping, prompt truncation
- test_episodic_memory_retention.py  — delete_old_episodes policy,
                                       future migration compatibility
- test_episodic_memory_min_score.py  — FTS5 min_score filtering, notes
                                       tier, resolve helper, has_fts5 property

Removes test_episodic_memory.py from GRANDFATHERED_FILES in file_size.py.
Updates two links in docs/rfcs/0017-pr-plan.md to point to the new module.
