+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-25  
-> **397** tracked files
+> **402** tracked files
 
 ## Summary
 
@@ -10,7 +10,7 @@
 | `internal/` | 93 |
 | `docs/` | 81 |
 | `agents/` | 77 |
-| `tests/` | 56 |
+| `tests/` | 61 |
 | `(root)` | 20 |
 | `scripts/` | 17 |
 | `.github/` | 15 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 144 |
+| `.py` | 149 |
 | `.md` | 97 |
 | `.go` | 90 |
 | `.yaml` | 25 |
@@ -487,7 +487,12 @@
 │           ├── test_defaults.py
 │           ├── test_dispatch_execute_actions.py
 │           ├── test_docker_fixes.py
-│           ├── test_episodic_memory.py
+│           ├── test_episodic_memory_compression.py
+│           ├── test_episodic_memory_core.py
+│           ├── test_episodic_memory_fts5.py
+│           ├── test_episodic_memory_min_score.py
+│           ├── test_episodic_memory_retention.py
+│           ├── test_episodic_memory_validation.py
 │           ├── test_event_dispatch_tick.py
 │           ├── test_llm_client.py
 │           ├── test_llm_persona_agent.py
